Optimize TypeRegistry::lookupName() from O(N) to O(1) (#6082)
|      Q       |   A
|------------- | -----------
| Type         | performance issue
| Fixed issues | n/a

#### Summary

We use a lot of custom types and migration from `Type::getName()` to
`TypeRegistry->lookupName()` discovered us the reverse lookup was very
slow (O(N)). This PR makes it O(1).

~Index is initialized lazily on the 1st reverse lookup.~ The type
registry is expected to not hold duplicate type instances and this PR
fixes it, in
https://github.com/doctrine/dbal/pull/6083#discussion_r1261812593 I was
requested to fix it in this PR instead of a separate one.