Apply the connection filter to indexes
When there are indexes that are not supported by doctrine present in
a database, they will consistently create "DROP" migrations.  By
applying the schema filter to the indexes one will be able to exclude
these indexes from future migrations preventing the manual work of
removing the DROP statement each time a migration is created.

Fixes: #5420