Generalize CustomIntrospectionTest (#6210)
|      Q       |   A
|------------- | -----------
| Type         | improvement
| Fixed issues | Follows #6189

#### Summary

This PR improves the test added in #6189 by allowing it to run on other
platforms than MySQL/MariaDB and by using non-deprecated APIs only. This
should make it easier to merge this test and later changes to it up to
the 4.0.x branch.

I've skipped Oracle though because I currently don't have a local Oracle
setup to test against.