Fix MariaDB list columns performance (#6202)
|      Q       |   A
|------------- | -----------
| Type         | bug
| Fixed issues | https://github.com/contao/contao/issues/6409

#### Summary

Starting with version 3.7.0 we noticed that the `listTableColumns()`
method got very slow on some systems, our migrate command went from
taking around 2 seconds with version 3.6 to 260 seconds with version3.7.

This is due to the `LEFT JOIN information_schema.CHECK_CONSTRAINTS` not
getting optimized by MariaDB which causes the database server to scan
every table of every database.

This pull request fixes that by using a subquery that is limited by
`WHERE CONSTRAINT_SCHEMA = $databaseName`. This brings the performance
back to where it was with version 3.6.