Promote private properties and methods to protected in `QueryBuilder`
QueryBuilder may be used in a decorator pattern, to enhance the
provided functionality. Private properties and methods makes it
nearly impossible to do the same job without copying all of the
private properties and methods and increasing the ongoing needed
maintenance burden.

This change promotes the private properties and method up to
protected accessibilty, which provides the same surface from
the outside but making frameworks and application lives easier
to provide enhancements with a minimal maintanence burden.