Use different toctrees for different sections
It will make clearer what category every article falls into. Also it
addresses an issue that results in the second section disappearing
whenever you click on any link.
It reveals classification issues in the existing docs. For instance, I
think architecture should go in the "Explanation" section, and that
there should probably be a tutorial and how-to section. If we decide to
do such moves, we should probably set up redirects, and I do not know
how to achieve that.