Switch to absolute paths
There is a piece of code in doctrine-website that adds leading slash,
and comes with a comment saying to remove that piece of code once all
sidebar files are fixed. It does so only for directories named
"reference", "tutorials" or "cookbook", which might explain why items
under explanation disappear once you navigate deeper in the tree.