Build temporary table name for SQLite from unqualified table name
In some situations, SqlitePlatform::getAlterTableSQL() relies on
a temporary table to store data from the original table while it
gets recreated. The name of this temporary table is built by
concatenating a prefix (__temp__) with the name of the original table.
When the original table name is foo.bar, there is an issue because
then the temporary table name becomes __temp__foo.bar which implies schema.table. Use the unqualified table name
so that the temporary table becomes __temp__bar.