Make all mapped database types case insensitive (#6321)
<!-- Fill in the relevant information below to help triage your pull
request. -->

|      Q       |   A
|------------- | -----------
| Type         | improvement
| Fixed issues | <!-- use #NUM format to reference an issue -->

#### Summary

<!-- Provide a summary of your change. -->
Currently, a custom mapping type has to return an array of lowercase
strings in `getMappedDatabaseTypes` or else Doctrine fails to retrieve
the added type as both `AbstractPlatform::getDoctrineTypeMapping` and
`AbstractPlatform::hasDoctrineTypeMappingFor` do a lowercase conversion
before the look-up. A devolper should not depend on internal
representation.

This PR addresses this by storing the mapped type in lowercase to
`AbstractPlatform#doctrineTypeMapping`. This is in line with
`AbstractPlatform::registerDoctrineTypeMapping`, which does a lowercase
conversion as well.