Fix SQL Server "extended property" SQL generation (#6353)
|      Q       |   A
|------------- | -----------
| Type         | bug
| Fixed issues | #4283

#### Summary


https://learn.microsoft.com/en-us/sql/relational-databases/system-stored-procedures/sp-addextendedproperty-transact-sql?view=sql-server-ver15
shows the `sp_addextendedproperty` accepts `sysname` datatype which is
subtype of string datatype and therefore the values must be escaped as
string literal instead of escaped identifier. This is evident also in
the
[examples](https://learn.microsoft.com/en-us/sql/relational-databases/system-stored-procedures/sp-addextendedproperty-transact-sql?view=sql-server-ver15#examples).

As the DBAL method arguments accepts possibly escaped identifier, we
unescape it before we escape the argument as string literal.