Use 3.8.x as a target for dependabot version updates
By default, Dependabot checks the default branch, and raises pull
requests against it. Since we are using a merge-up workflow, we should
tell it to target 3.8.x. This is not great because it will mean we will
have to maintain that setting for as long as we maintain the old patch
branch, but I do not think there is a better solution.