Fix MariaDB fetching of default table character-set (#6361)
From MariaDB-10.10.1, where uca1400 was added, the
information_schema.COLLATION_CHARACTER_SET_APPLICABILITY was
extended to have FULL_COLLATION_NAME which corresponds to
the information_schema.TABLES.TABLE_COLLATION value.

Executable comment syntax is used to limited to the applicable
versions.

To preserve compatibility with older MariaDB versions, and
MySQL versions where the previous COLLATION_NAME was the match
is left as a JOIN critieria. In new MariaDB versions this won't
result in an extra row match.

Closes: #6361