Fix MariaDB fetching of default table character-set (#6361) (#6425)
<!-- Fill in the relevant information below to help triage your pull
request. -->

|      Q       |   A
|------------- | -----------
| Type         | bug
| Fixed issues | #6361

#### Summary

From MariaDB-10.10.1, where uca1400 was added, the
information_schema.COLLATION_CHARACTER_SET_APPLICABILITY was extended to
have FULL_COLLATION_NAME which corresponds to the
information_schema.TABLES.TABLE_COLLATION value.

Executable comment syntax is used to limited to the applicable versions.

To preserve compatibility with older MariaDB versions, and MySQL
versions where the previous COLLATION_NAME was the match is left as a
JOIN criteria. In new MariaDB versions this won't result in an extra row
match.

Closes: #6361


The lack of this fix did case the CI test to fail with a MariaDB-11.0+
container:


```
1) Doctrine\DBAL\Tests\Functional\Schema\MySQLSchemaManagerTest::testEnsureTableWithoutOptionsAreReflectedInMetadata
Undefined array key "engine"

/home/runner/work/dbal/dbal/src/Schema/Table.php:927
/home/runner/work/dbal/dbal/tests/Functional/Schema/MySQLSchemaManagerTest.php:550
```


With this fix all version of MySQL and MariaDB (even those new ones soon
to be in a different PR) will pass like:
https://github.com/grooverdan/dbal/actions/runs/9412110648/job/25926481038