CI: Update MariaDB versions (#6426)
<!-- Fill in the relevant information below to help triage your pull
request. -->

|      Q       |   A
|------------- | -----------
| Type         | improvement
| Fixed issues | N/A <!-- use #NUM format to reference an issue -->

#### Summary

Improve the CI to cover newer versions of MariaDB 11.4+.

Also cover the quay.io/mariadb-foundation/mariadb-devel images from
https://quay.io/repository/mariadb-foundation/mariadb-devel?tab=tags
where they are updated with completed, reviewed and tested changes in
the MariaDB server main branch.

These are completed features/bug fixes in the next server version. As
such this is a good thing to test before the users get the final
product.

ref:
https://mariadb.org/new-service-quay-io-mariadb-foundation-mariadb-devel/

The test failures here are fixed by #6425.

<!-- Provide a summary of your change. -->