Fix update/delete aliases in documentation (#6470)
<!-- Fill in the relevant information below to help triage your pull
request. -->

|      Q       |   A
|------------- | -----------
| Type         | bug/improvement
| Fixed issues | Incorrect & inconsistent QueryBuilder documentation for
aliases in update/delete

#### Summary

In #6394 the documentation was partially updated to remove aliases for
the documentation for the update method, as the separate parameter for
it was removed in the upgrade from 3.x to 4.x.

Inline aliasing is still possible, and I suspect this was the reason the
extra parameter was removed. ( `->update('users alias1', 'alias2')`)?

This PR readds the alias to the documentation for the occurences in the
linked PR, and also updates the other occurences in docblocks + the
online documentation.