CI: Add MySQL 9, reduce test matrix (#6462)
MySQL 9 has been released and 8.4 has been declared to be the LTS
release. I've added MySQL 9 to the test matrix and took the opportunity
to shrink the test matrix a little. Our CI is othen blocked because of
too many concurrent jobs, so we might not want to test all possible
permutations of PHP and MySQL versions.

Taking into account that we're going to maintain the 3.x branch for a
little longer, we need a strategy for not letting the test matrix grow
exponentially with every future PHP or DBMS release.

My proposal for MySQL:

* Test all supported MySQL versions (currently: 5.7, 8.0, 8.4, 9.0) with
the latest PHP version (currently 8.3).
* Test the oldest supported PHP release (currently 7.4) against one
release only (MySQL 8.0) only.

All other PHP releases are tested against SQLite already.

If this proposal is accepted, I'd like to work out a similar strategy
for the other supported DBMS. I think we should also document that
strategy properly then.