Make the schema manager aware of the disabling of type comments
Without this change, the schema comparison will not remove existing type
comments when disabling them because it will still silently remove that
part of the comment during introspection (and so won't detect an
expected change between the expected empty comment and the introspected
comment).
Disabling that stripping logic when disabling type comments corresponds
to the behavior of DBAL 4 regarding comments, which is exactly what the
configuration option is about.