Handle cached result column names and rows separately (#6504)
|      Q       |   A
|------------- | -----------
| Type         | bug

#### Summary

Prior to #6428, there were two problems with caching results by design:

1. If the result contains no rows, its cached version won't return the
number of columns.
2. If the result contains columns with the same name (e.g. `SELECT a.id,
b.id FROM a, b`), they will clash even if fetched in the numeric mode.

See https://github.com/doctrine/dbal/pull/6428#issuecomment-2241896892
for reference.

The solution is:
1. Handle column names and rows in the cache result separately. This
way, the column names are available even if there is no data.
2. Store rows as tuples (lists) instead of maps (associative arrays).
This way, even if the result contains multiple columns with the same
name, they can be correctly fetched with `fetchNumeric()`.

**Additional notes**:
1. The changes in `TestUtil::generateResultSetQuery()` were necessary
for the integration testing of the changes in `ArrayResult`. However,
later, I realized that the modified code is database-agnostic, so
running it in integration with each database platform would effectively
test the same code path and would be redundant. I replaced the
integration tests with the unit ones but left the change in `TestUtil`
as it makes the calling code cleaner.
2. `ArrayStatementTest` was renamed to `ArrayResultTest`, which Git
doesn't seem to recognize due to the amount of changes.

**TODO**:
- [x] Decide on the release version. This pull request changes the cache
format, so the users will have to clear their cache during upgrade. I
wouldn't consider it a BC break. We could also migrate the cache at
runtime, but given that this cache can be cleared, I don't think it is
worth the effort.
- [x] Add upgrade documentation.