Invalidate old query cache format (#6510)
|      Q       |   A
|------------- | -----------
| Type         | improvement
| Fixed issues | Follows #6504

#### Summary

I'm changing the query cache format once again: Instead of storing a
tuple of arrays, I'm now storing the whole `ArrayResult` object. This
way, we can easily detect if we support the data returned from the cache
and we can handle future cache format changes in the `__unserialize()`
methods.

After #6504, we would run into errors if an app would attempt to us a
cache written with DBAL 4.1. With my change, the old cache is ignored
and the app would behave as if it had encountered a cache miss instead.

I've also added tests covering the serialization of `ArrayResult`
objects.