PDO: Raise a proper exception if user or password is false (#6513)
|      Q       |   A
|------------- | -----------
| Type         | bug
| Fixed issues | #6512 

#### Summary

If `false` (or anything that is not a string) is passed as `user` or
`password`, we run into a `TypeError` because we pass that value as-is
to the constructor of PDO. This started to happen after we enabled
strict types on our driver classes in 4.0.

On 3.9, `false` would implicitly be cast to an empty string which is
either desired or leads to more obscure connection errors. We could
restore the behavior of 3.9 by adding explicit type casts to the two
parameters. But since we don't document `false` as a valid value for
either parameter, my preference would indeed be raising an exception.