Modify test
The test as previously modified succeeded without the stream restore
action (comment out the fseek call in src/Driver/PDO/Statement.php).

The expectation with the fix is that the resource sent to bindValue is
read during the first call to execute. This is the point the stream
pointer is moved and needs to be removed. For PDO SQLite, additional
calls execute (without rebinding the stream value) do not read the
stream a second time.

If this needs to be applied for subsequent calls to execute, then
bindValue needs to forward the provided $param to trackParamResource
for a paramResources index, and paramResources should not be cleared
at the end of restoreResourceOffsets. I need a ci pass to see if
this needed.

If the test needs to run without the ftell check, then the statement
needs to be recreated and/or bindValue called a second time on the
stream before executing the statement. This will force a second read
of the stream, whereas currently it is only read one time.