DB2 execute with large object fails second call
An execute* function with a bound large object parameter can be called
once, but not twice.

Execute is clearing the lob collection, but is not removing the
handles this data is written to from the bound parameters. Clear the
parameters instead and leave the lob collection intact so that it can
be used in subsequent calls to repopulate new handles.

Note that the resources in the lob collection have already been
restored to their original resource positions.