fix(PgSQL): NULL handling when binding values
PostgreSQL allows 3 values in a boolean column: TRUE, FALSE and NULL.

Previously, when binding `NULL` to a boolean column, it passed the value
as FALSE to the database. It prevents having NULL values in the column,
when using Doctrine DBAL.

It fixes this issue passing `NULL` forward to the database.

The test has been modified to test this scenario.